omega-logger/requirements.txt
paho-mqtt>=2.0.0
pymodbus>=3.6.0
pyserial>=3.5
pyyaml>=6.0
